Move HPA E2E tests to a separate package.
